+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/page_blob_client.hpp
@@ -209,53 +209,6 @@ namespace Azure { namespace Storage { namespace Blobs {
         const Azure::Core::Context& context = Azure::Core::Context()) const;
 
     /**
-     * @brief Returns the list of valid page ranges for a page blob or snapshot of a page blob.
-     *
-     * @param options Optional parameters to execute this function.
-     * @param context Context for cancelling long running operations.
-     * @return A GetPageRangesResult describing the valid page ranges for this blob.
-     */
-    Azure::Response<Models::GetPageRangesResult> GetPageRanges(
-        const GetPageRangesOptions& options = GetPageRangesOptions(),
-        const Azure::Core::Context& context = Azure::Core::Context()) const;
-
-    /**
-     * @brief Returns the list of page ranges that differ between a previous snapshot and this page
-     * blob. Changes include both updated and cleared pages.
-     *
-     * @param previousSnapshot Specifies that the response will contain only pages that were changed
-     * between target blob and previous snapshot. Changed pages include both updated and cleared
-     * pages. The target blob may be a snapshot, as long as the snapshot specified by
-     * previousSnapshot is the older of the two.
-     * @param options Optional parameters to execute this function.
-     * @param context Context for cancelling long running operations.
-     * @return A GetPageRangesResult describing the valid page ranges for this blob.
-     */
-    Azure::Response<Models::GetPageRangesResult> GetPageRangesDiff(
-        const std::string& previousSnapshot,
-        const GetPageRangesOptions& options = GetPageRangesOptions(),
-        const Azure::Core::Context& context = Azure::Core::Context()) const;
-
-    /**
-     * @brief Returns the list of page ranges that differ between a previous snapshot url and this
-     * page blob. Changes include both updated and cleared pages. This API only works with managed
-     * disk storage accounts.
-     *
-     * @param previousSnapshotUrl This parameter only works with managed disk storage accounts.
-     * Specifies that the response will contain only pages that were changed between target blob and
-     * previous snapshot. Changed pages include both updated and cleared pages. The target blob may
-     * be a snapshot, as long as the snapshot specified by previousSnapshotUrl is the older of the
-     * two.
-     * @param options Optional parameters to execute this function.
-     * @param context Context for cancelling long running operations.
-     * @return A GetPageRangesResult describing the valid page ranges for this blob.
-     */
-    Azure::Response<Models::GetPageRangesResult> GetManagedDiskPageRangesDiff(
-        const std::string& previousSnapshotUrl,
-        const GetPageRangesOptions& options = GetPageRangesOptions(),
-        const Azure::Core::Context& context = Azure::Core::Context()) const;
-
-    /**
      * @brief Starts copying a snapshot of the sourceUri page blob to this page blob. The snapshot
      * is copied such that only the differential changes between the previously copied snapshot
      * are transferred to the destination. The copied snapshots are complete copies of the original